removed all duplicated agents configurations

+++ AgentMesh.Application/Configuration/AgentFlatConfigurationRecord.cs
@@ -10,8 +10,8 @@
         public readonly string ProviderModelName { get; init; }
 
         public readonly string LLMClass { get; init; }
-        public readonly double LLMClassCostPerMillionInputTokens { get; init; }
-        public readonly double LLMClassCostPerMillionOutputTokens { get; init; }
+        public readonly decimal LLMClassCostPerMillionInputTokens { get; init; }
+        public readonly decimal LLMClassCostPerMillionOutputTokens { get; init; }
         
         public readonly string Temperature { get; init; }
         public readonly string SystemPrompt { get; init; }
